ocrapi/requirements.txt
Flask==3.0.3
gunicorn==22.0.0
google-cloud-vision==3.7.3
python-dotenv==1.0.1
Pillow==10.4.0
